requirements.txt
discord.py>=1.3.3
requests>=2.24
lxml>=4.5.1
